+++ Code/ManualGPIO.h
@@ -7,6 +7,6 @@
 int stop;
 void pinMode(int pin, short mode);
 void digitalWrite(int pin,short value);
-int digitalRead(short pin);
+int digitalRead(int pin);
 void blink(short pin, double frequency,double duration);
 #endif //IMAGETEST_MANUALGPIO_H
